+++ tests/cxx_tests.cmake
@@ -28,6 +28,7 @@ add_executable(${test_target}
 target_include_directories(${test_target}
     PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}
+    ${OpenTLV_SOURCE_DIR}/tests
 )
 
 target_link_libraries(${test_target} PRIVATE